{{cookiecutter.repo_name}}/requirements.txt
apispec==0.39.0
attrs==17.4.0
botocore==1.12.86
chalice==1.6.2
click==6.7
docutils==0.14
enum-compat==0.0.2
invoke==1.2.0
jmespath==0.9.3
marshmallow==2.18.0
python-dateutil==2.7.5
PyYAML==3.13
six==1.12.0
typing==3.6.4
urllib3==1.24.1
